docs(core): added common solutions page
Creates a feature branch and file for common use cases solved in a
"nestic" fashion, following NestJS design patterns and using the
appropriate NestJS functionality. This contribution aims to allow more
experienced nest developers to share their expertise with those new to
the framework, so they can tackle these issues and get practical
examples of how the different pieces of the documentation can be
applied to their code.